remove TableViewController
